Refactor #264: Cleaned up server.js, removed broken connectDB block, and fixed missing setupGracefulShutdown import
